fix restarting after refresh
